// src/chess/gamelogic/pieces/types.ts
export enum PieceName {
  PAWN = "PAWN",
  KING = "KING",
  BISHOP = "BISHOP",
  KNIGHT = "KNIGHT",
  QUEEN = "QUEEN",
  ROOK = "ROOK",
}

export enum Direction {
  UP = "UP",
  UP_RIGHT = "UP_RIGHT",
  RIGHT = "RIGHT",
  DOWN_RIGHT = "DOWN_RIGHT",
  DOWN = "DOWN",
  DOWN_LEFT = "DOWN_LEFT",
  LEFT = "LEFT",
  UP_LEFT = "UP_LEFT",
}

export enum PieceStateName {
  MOVING = "MOVING",
  IDLE = "IDLE",
  MOVED = "MOVED",
}